⏪️ Reverted `App.vue` and moved 3d model tests to the design page
